Added Command Line Interface, clients can view eachothers images, send requests, accept requests, accept view counts
Added another client, the program flows as a CLI and each user can add, delete, view all images, or view a certain users images. Clients can now also send a request to another client with the number of views they want for a certain image. Then the other client receives this notification and can accept or deny, this is then reflected to the client and to the directory.json.

it works so far through with 2 clients each having the others target port defined. The clients communicate through TCP.