Function for clients to edit who has access to their images (besides accepting requests), placeholder image for preview, error checking for requests before sending them.)
Added function to allow a user to view the images they own and edit who has access and how many views. Added error checking to make sure the client is making a request for an image that exists and isnt owned by them. Added placeholder image for previewing the directory when its empty.